Refactor code

+++ lib/gpx/src/addWaypoint.cpp
@@ -8,7 +8,7 @@
 
 #include "addWaypoint.hpp"
 
-extern const int SD_CS;
+extern const uint8_t SD_CS;
 
 wayPoint addWpt = {0, 0, 0, (char *)"", (char *)"", (char *)"", (char *)"", (char *)"", (char *)"", 0, 0, 0, 0};
 wayPoint loadWpt = {0, 0, 0, (char *)"", (char *)"", (char *)"", (char *)"", (char *)"", (char *)"", 0, 0, 0, 0};
